Add compile options and Rot and Over intructions
Rot rotates the stack: The top element moves to the bottom.
Over adds a copy of the top element to the bottom of the stack.

I tried implementing multiply, but couldn't with the previous
instruction set. I think these are what I need to do it. They are
included operations in Forth.